requirements_dev.txt
homeassistant==2025.12.4
aiohttp==3.13.2
pyserial
modbus-tk